Move RuntimeError logic into sys
