+++ lib/api/src/sys/instance.rs
@@ -29,8 +29,8 @@ mod send_test {
 impl From<wasmer_compiler::InstantiationError> for InstantiationError {
     fn from(other: wasmer_compiler::InstantiationError) -> Self {
         match other {
-            wasmer_compiler::InstantiationError::Link(e) => Self::Link(e),
-            wasmer_compiler::InstantiationError::Start(e) => Self::Start(e),
+            wasmer_compiler::InstantiationError::Link(e) => Self::Link(e.into()),
+            wasmer_compiler::InstantiationError::Start(e) => Self::Start(e.into()),
             wasmer_compiler::InstantiationError::CpuFeature(e) => Self::CpuFeature(e),
         }
     }
