+++ lib/compiler/src/engine/error.rs
@@ -1,8 +1,8 @@
 //! The WebAssembly possible errors
-#[cfg(not(target_arch = "wasm32"))]
-use crate::engine::trap::RuntimeError;
 use thiserror::Error;
 pub use wasmer_types::{DeserializeError, ImportError, SerializeError};
+#[cfg(not(target_arch = "wasm32"))]
+use wasmer_vm::Trap;
 
 /// The WebAssembly.LinkError object indicates an error during
 /// module instantiation (besides traps from the start function).
@@ -19,8 +19,8 @@ pub enum LinkError {
 
     #[cfg(not(target_arch = "wasm32"))]
     /// A trap ocurred during linking.
-    #[error("RuntimeError occurred during linking: {0}")]
-    Trap(#[source] RuntimeError),
+    #[error("Trap occurred during linking: {0}")]
+    Trap(#[source] Trap),
 
     /// Insufficient resources available for linking.
     #[error("Insufficient resources: {0}")]
@@ -48,5 +48,5 @@ pub enum InstantiationError {
     /// A runtime error occured while invoking the start function
     #[cfg(not(target_arch = "wasm32"))]
     #[error(transparent)]
-    Start(RuntimeError),
+    Start(Trap),
 }
